RIA-7615 Changes after rebase 1
