+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/detentionengagementteam/DetentionEngagementTeamAppealFeeDuePersonalisation.java
@@ -1,13 +1,13 @@
 package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.detentionengagementteam;
 
 import static java.util.Objects.requireNonNull;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.DETENTION_FACILITY;
 import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DocumentTag.INTERNAL_APPEAL_FEE_DUE_LETTER;
 import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.utils.AsylumCaseUtils.getLetterForNotification;
 
-import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
-import java.util.*;
+import java.util.Map;
+import java.util.Set;
+import com.google.common.collect.ImmutableMap;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Value;
@@ -19,45 +19,41 @@ import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.DetEmailService
 import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.DocumentDownloadClient;
 import uk.gov.service.notify.NotificationClientException;
 
-@Service
+
 @Slf4j
+@Service
 public class DetentionEngagementTeamAppealFeeDuePersonalisation implements EmailWithLinkNotificationPersonalisation {
 
     private final String internalDetainedAppealFeeDueTemplateId;
     private final DocumentDownloadClient documentDownloadClient;
-    private final String detainedPrefix;
     private final DetEmailService detEmailService;
+    private String subjectPrefix;
 
     public DetentionEngagementTeamAppealFeeDuePersonalisation(
-        @Value("${govnotify.template.appealFeeIsDue.detentionEngagementTeam.nonAda.email}") String templateIdDetained,
-        @Value("${govnotify.emailPrefix.nonAdaInPerson}") String detainedPrefix,
-        DetEmailService detEmailService,
-        DocumentDownloadClient documentDownloadClient
+            @Value("${govnotify.template.pendingPaymentEaHuEu.detentionEngagementTeam.email}") String internalDetainedAppealFeeDueTemplateId,
+            DetEmailService detEmailService,
+            DocumentDownloadClient documentDownloadClient,
+            @Value("${govnotify.emailPrefix.nonAdaInPerson}") String subjectPrefix
     ) {
-        this.internalDetainedAppealFeeDueTemplateId = templateIdDetained;
-        this.detainedPrefix = detainedPrefix;
+        this.internalDetainedAppealFeeDueTemplateId = internalDetainedAppealFeeDueTemplateId;
         this.detEmailService = detEmailService;
         this.documentDownloadClient = documentDownloadClient;
+        this.subjectPrefix = subjectPrefix;
     }
 
     @Override
-    public String getTemplateId(AsylumCase asylumCase) {
+    public String getTemplateId() {
         return internalDetainedAppealFeeDueTemplateId;
     }
 
     @Override
     public Set<String> getRecipientsList(AsylumCase asylumCase) {
-        Optional<String> detentionFacility = asylumCase.read(DETENTION_FACILITY, String.class);
-        if (detentionFacility.isEmpty() || detentionFacility.get().equals("other")) {
-            return Collections.emptySet();
-        }
-
-        return Collections.singleton(detEmailService.getDetEmailAddress(asylumCase));
+        return detEmailService.getRecipientsList(asylumCase);
     }
 
     @Override
     public String getReferenceId(Long caseId) {
-        return caseId + "_INTERNAL_DET_APPEAL_FEE_DUE_EMAIL";
+        return caseId + "_INTERNAL_DETAINED_APPEAL_FEE_DUE_DET";
     }
 
     @Override
@@ -65,22 +61,22 @@ public class DetentionEngagementTeamAppealFeeDuePersonalisation implements Email
         requireNonNull(asylumCase, "asylumCase must not be null");
 
         return ImmutableMap
-            .<String, Object>builder()
-            .put("subjectPrefix", detainedPrefix)
-            .put("appealReferenceNumber", asylumCase.read(AsylumCaseDefinition.APPEAL_REFERENCE_NUMBER, String.class).orElse(""))
-            .put("homeOfficeReferenceNumber", asylumCase.read(AsylumCaseDefinition.HOME_OFFICE_REFERENCE_NUMBER, String.class).orElse(""))
-            .put("appellantGivenNames", asylumCase.read(AsylumCaseDefinition.APPELLANT_GIVEN_NAMES, String.class).orElse(""))
-            .put("appellantFamilyName", asylumCase.read(AsylumCaseDefinition.APPELLANT_FAMILY_NAME, String.class).orElse(""))
-            .put("documentLink", getAppealFeeIsDueDocumentInJsonObject(asylumCase))
-            .build();
+                .<String, Object>builder()
+                .put("subjectPrefix", subjectPrefix)
+                .put("appealReferenceNumber", asylumCase.read(AsylumCaseDefinition.APPEAL_REFERENCE_NUMBER, String.class).orElse(""))
+                .put("homeOfficeReferenceNumber", asylumCase.read(AsylumCaseDefinition.HOME_OFFICE_REFERENCE_NUMBER, String.class).orElse(""))
+                .put("appellantGivenNames", asylumCase.read(AsylumCaseDefinition.APPELLANT_GIVEN_NAMES, String.class).orElse(""))
+                .put("appellantFamilyName", asylumCase.read(AsylumCaseDefinition.APPELLANT_FAMILY_NAME, String.class).orElse(""))
+                .put("documentLink", getInternalDetainedAppealFeeDueLetterInJsonObject(asylumCase))
+                .build();
     }
 
-    private JSONObject getAppealFeeIsDueDocumentInJsonObject(AsylumCase asylumCase) {
+    private JSONObject getInternalDetainedAppealFeeDueLetterInJsonObject(AsylumCase asylumCase) {
         try {
             return documentDownloadClient.getJsonObjectFromDocument(getLetterForNotification(asylumCase, INTERNAL_APPEAL_FEE_DUE_LETTER));
         } catch (IOException | NotificationClientException e) {
-            log.error("Failed to get Internal 'Appeal fee is due' Letter in compatible format", e);
-            throw new IllegalStateException("Failed to get Internal 'Appeal fee is due' Letter in compatible format");
+            log.error("Failed to get Internal detained appeal fee due letter in compatible format", e);
+            throw new IllegalStateException("Failed to get Internal detained appeal fee due letter in compatible format");
         }
     }
 }
