+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/infrastructure/config/NotificationGeneratorConfiguration.java
@@ -4117,13 +4117,13 @@ public class NotificationGeneratorConfiguration {
 
     @Bean("internalDetainedAppealFeeDueNotificationGenerator")
     public List<NotificationGenerator> internalDetainedAppealFeeDueNotificationGenerator(
-            DetentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation,
+            DetentionEngagementTeamAppealFeeDuePersonalisation detentionEngagementTeamAppealFeeDuePersonalisation,
             GovNotifyNotificationSender notificationSender,
             NotificationIdAppender notificationIdAppender) {
 
         return Arrays.asList(
                 new EmailWithLinkNotificationGenerator(
-                        newArrayList(Collections.singleton(detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation)),
+                        newArrayList(Collections.singleton(detentionEngagementTeamAppealFeeDuePersonalisation)),
                         notificationSender,
                         notificationIdAppender
                 )
