--- src/test/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/detentionengagementteam/DetentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisationTest.java
@@ -1,152 +1,0 @@
-package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.detentionengagementteam;
-
-import static com.google.common.collect.Lists.newArrayList;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.TestUtils.compareStringsAndJsonObjects;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.TestUtils.getDocumentWithMetadata;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.*;
-
-import java.io.IOException;
-import java.util.Collections;
-import java.util.Map;
-import java.util.Optional;
-import com.google.common.collect.ImmutableMap;
-import org.json.JSONObject;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.mockito.junit.jupiter.MockitoSettings;
-import org.mockito.quality.Strictness;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCase;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DocumentTag;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DocumentWithMetadata;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.field.IdValue;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.DetEmailService;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.DocumentDownloadClient;
-import uk.gov.service.notify.NotificationClientException;
-
-
-@ExtendWith(MockitoExtension.class)
-@MockitoSettings(strictness = Strictness.LENIENT)
-public class DetentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisationTest {
-    @Mock
-    AsylumCase asylumCase;
-    @Mock
-    DetEmailService detEmailService;
-    @Mock
-    JSONObject jsonDocument;
-    @Mock
-    DocumentDownloadClient documentDownloadClient;
-
-    private final String templateId = "someTemplateId";
-    private final String detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisationReferenceId = "_INTERNAL_DETAINED_APPEAL_FEE_DUE_DET";
-    private final String detEmailAddress = "some@example.com";
-    private final String appealReferenceNumber = "someReferenceNumber";
-    private final String homeOfficeReferenceNumber = "1234-1234-1234-1234";
-    private final String appellantGivenNames = "someAppellantGivenNames";
-    private final String appellantFamilyName = "someAppellantFamilyName";
-    private final String subjectPrefix = "IAFT - SERVE IN PERSON";
-    DocumentWithMetadata internalDetainedAppealFeeDueLetter = getDocumentWithMetadata(
-            "1", "internal-detained-payment-due-no-remission-letter", "some other desc", DocumentTag.INTERNAL_APPEAL_FEE_DUE_LETTER);
-    IdValue<DocumentWithMetadata> internalDetainedAppealFeeDueLetterId = new IdValue<>("1", internalDetainedAppealFeeDueLetter);
-
-    private DetentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation;
-
-
-    @BeforeEach
-    public void setUp() throws NotificationClientException, IOException {
-
-        when(asylumCase.read(APPEAL_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(appealReferenceNumber));
-        when(asylumCase.read(APPELLANT_GIVEN_NAMES, String.class)).thenReturn(Optional.of(appellantGivenNames));
-        when(asylumCase.read(APPELLANT_FAMILY_NAME, String.class)).thenReturn(Optional.of(appellantFamilyName));
-        when(asylumCase.read(HOME_OFFICE_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(homeOfficeReferenceNumber));
-
-        when(asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS)).thenReturn(Optional.of(newArrayList(internalDetainedAppealFeeDueLetterId)));
-        when(documentDownloadClient.getJsonObjectFromDocument(internalDetainedAppealFeeDueLetter)).thenReturn(jsonDocument);
-
-        detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation =
-                new DetentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation(
-                        templateId,
-                        detEmailService,
-                        documentDownloadClient,
-                        subjectPrefix
-                );
-    }
-
-    @Test
-    public void should_return_given_template_id() {
-        assertEquals(templateId, detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getTemplateId());
-    }
-
-    @Test
-    public void should_return_given_reference_id() {
-        Long caseId = 12345L;
-        assertEquals(caseId + detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisationReferenceId,
-                detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getReferenceId(caseId));
-    }
-
-    @Test
-    public void should_return_given_email_address_from_asylum_case() {
-        when(asylumCase.read(DETENTION_FACILITY, String.class)).thenReturn(Optional.of("immigrationRemovalCentre"));
-        when(detEmailService.getRecipientsList(asylumCase)).thenReturn(Collections.singleton(detEmailAddress));
-
-        assertTrue(
-                detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getRecipientsList(asylumCase).contains(detEmailAddress));
-    }
-
-    @Test
-    public void should_return_empty_set_email_address_from_asylum_case_no_detention_facility() {
-        when(asylumCase.read(DETENTION_FACILITY, String.class)).thenReturn(Optional.empty());
-        assertEquals(Collections.emptySet(), detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getRecipientsList(asylumCase));
-    }
-
-    @Test
-    public void should_return_empty_set_email_address_from_asylum_case_other_detention_facility() {
-        when(asylumCase.read(DETENTION_FACILITY, String.class)).thenReturn(Optional.of("other"));
-        assertEquals(Collections.emptySet(), detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getRecipientsList(asylumCase));
-    }
-
-    @Test
-    public void should_throw_exception_on_personalisation_when_case_is_null() {
-
-        assertThatThrownBy(
-                () -> detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getPersonalisationForLink((AsylumCase) null))
-                .isExactlyInstanceOf(NullPointerException.class)
-                .hasMessage("asylumCase must not be null");
-    }
-
-    @Test
-    public void should_throw_exception_on_personalisation_when_internal_detained_appeal_fee_due_document_is_missing() {
-        when(asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(
-                () -> detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getPersonalisationForLink(asylumCase))
-                .isExactlyInstanceOf(IllegalStateException.class)
-                .hasMessage("internalAppealFeeDueLetter document not available");
-    }
-
-    @Test
-    public void should_return_personalisation_when_all_information_given_refused() {
-
-        final Map<String, Object> expectedPersonalisation =
-                ImmutableMap
-                        .<String, Object>builder()
-                        .put("subjectPrefix", subjectPrefix)
-                        .put("appealReferenceNumber", appealReferenceNumber)
-                        .put("homeOfficeReferenceNumber", homeOfficeReferenceNumber)
-                        .put("appellantGivenNames", appellantGivenNames)
-                        .put("appellantFamilyName", appellantFamilyName)
-                        .put("documentLink", jsonDocument)
-                        .build();
-
-        Map<String, Object> actualPersonalisation =
-                detentionEngagementTeamAppealFeeDueNoRemissionHuEaEuPersonalisation.getPersonalisationForLink(asylumCase);
-
-        assertTrue(compareStringsAndJsonObjects(expectedPersonalisation, actualPersonalisation));
-    }
-}
