+++ src/test/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/detentionengagementteam/DetentionEngagementTeamAppealFeeDuePersonalisationTest.java
@@ -9,11 +9,11 @@ import static uk.gov.hmcts.reform.iacasenotificationsapi.TestUtils.compareString
 import static uk.gov.hmcts.reform.iacasenotificationsapi.TestUtils.getDocumentWithMetadata;
 import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.*;
 
-import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
+import com.google.common.collect.ImmutableMap;
 import org.json.JSONObject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -33,7 +33,6 @@ import uk.gov.service.notify.NotificationClientException;
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
 public class DetentionEngagementTeamAppealFeeDuePersonalisationTest {
-
     @Mock
     AsylumCase asylumCase;
     @Mock
@@ -43,20 +42,20 @@ public class DetentionEngagementTeamAppealFeeDuePersonalisationTest {
     @Mock
     DocumentDownloadClient documentDownloadClient;
 
-    private final String appealFeeDueTemplateId = "templateId";
-    private final String appealFeeDueReferenceId = "_INTERNAL_DET_APPEAL_FEE_DUE_EMAIL";
-    private final String detainedPrefix = "IAFT - SERVE IN PERSON";
-    private final String ircName = "Tinsley House";
+    private final String templateId = "someTemplateId";
+    private final String detentionEngagementTeamAppealFeeDuePersonalisationReferenceId = "_INTERNAL_DETAINED_APPEAL_FEE_DUE_DET";
     private final String detEmailAddress = "some@example.com";
     private final String appealReferenceNumber = "someReferenceNumber";
     private final String homeOfficeReferenceNumber = "1234-1234-1234-1234";
     private final String appellantGivenNames = "someAppellantGivenNames";
     private final String appellantFamilyName = "someAppellantFamilyName";
-    DocumentWithMetadata appealFeeDueLetterDoc = getDocumentWithMetadata(
-        "1", "Home Office directed to upload bundle", "some other desc", DocumentTag.INTERNAL_APPEAL_FEE_DUE_LETTER);
-    IdValue<DocumentWithMetadata> appealFeeDueLetterDocId = new IdValue<>("1", appealFeeDueLetterDoc);
-    private DetentionEngagementTeamAppealFeeDuePersonalisation
-        detentionEngagementTeamAppealFeeDuePersonalisation;
+    private final String subjectPrefix = "IAFT - SERVE IN PERSON";
+    DocumentWithMetadata internalDetainedAppealFeeDueLetter = getDocumentWithMetadata(
+        "1", "internal-detained-payment-due-no-remission-letter", "some other desc", DocumentTag.INTERNAL_APPEAL_FEE_DUE_LETTER);
+    IdValue<DocumentWithMetadata> internalDetainedAppealFeeDueLetterId = new IdValue<>("1", internalDetainedAppealFeeDueLetter);
+
+    private DetentionEngagementTeamAppealFeeDuePersonalisation detentionEngagementTeamAppealFeeDuePersonalisation;
+
 
     @BeforeEach
     public void setUp() throws NotificationClientException, IOException {
@@ -65,41 +64,35 @@ public class DetentionEngagementTeamAppealFeeDuePersonalisationTest {
         when(asylumCase.read(APPELLANT_GIVEN_NAMES, String.class)).thenReturn(Optional.of(appellantGivenNames));
         when(asylumCase.read(APPELLANT_FAMILY_NAME, String.class)).thenReturn(Optional.of(appellantFamilyName));
         when(asylumCase.read(HOME_OFFICE_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(homeOfficeReferenceNumber));
-        when(asylumCase.read(IRC_NAME, String.class)).thenReturn(Optional.of(ircName));
-        when(detEmailService.getDetEmailAddress(asylumCase)).thenReturn(detEmailAddress);
 
-        when(asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS)).thenReturn(Optional.of(newArrayList(appealFeeDueLetterDocId)));
-        when(documentDownloadClient.getJsonObjectFromDocument(appealFeeDueLetterDoc)).thenReturn(jsonDocument);
+        when(asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS)).thenReturn(Optional.of(newArrayList(internalDetainedAppealFeeDueLetterId)));
+        when(documentDownloadClient.getJsonObjectFromDocument(internalDetainedAppealFeeDueLetter)).thenReturn(jsonDocument);
 
         detentionEngagementTeamAppealFeeDuePersonalisation =
             new DetentionEngagementTeamAppealFeeDuePersonalisation(
-                appealFeeDueTemplateId,
-                detainedPrefix,
+                templateId,
                 detEmailService,
-                documentDownloadClient
+                documentDownloadClient,
+                subjectPrefix
             );
     }
 
-
     @Test
-    public void should_return_given_template_id_detained() {
-        assertEquals(
-            appealFeeDueTemplateId,
-            detentionEngagementTeamAppealFeeDuePersonalisation.getTemplateId(asylumCase)
-        );
+    public void should_return_given_template_id() {
+        assertEquals(templateId, detentionEngagementTeamAppealFeeDuePersonalisation.getTemplateId());
     }
 
     @Test
     public void should_return_given_reference_id() {
         Long caseId = 12345L;
-        assertEquals(caseId + appealFeeDueReferenceId,
+        assertEquals(caseId + detentionEngagementTeamAppealFeeDuePersonalisationReferenceId,
             detentionEngagementTeamAppealFeeDuePersonalisation.getReferenceId(caseId));
     }
 
     @Test
     public void should_return_given_email_address_from_asylum_case() {
         when(asylumCase.read(DETENTION_FACILITY, String.class)).thenReturn(Optional.of("immigrationRemovalCentre"));
-        when(detEmailService.getDetEmailAddress(asylumCase)).thenReturn(detEmailAddress);
+        when(detEmailService.getRecipientsList(asylumCase)).thenReturn(Collections.singleton(detEmailAddress));
 
         assertTrue(
             detentionEngagementTeamAppealFeeDuePersonalisation.getRecipientsList(asylumCase).contains(detEmailAddress));
@@ -127,11 +120,13 @@ public class DetentionEngagementTeamAppealFeeDuePersonalisationTest {
     }
 
     @Test
-    public void should_throw_exception_when_appeal_fee_is_due_letter_is_empty() {
+    public void should_throw_exception_on_personalisation_when_internal_detained_appeal_fee_due_document_is_missing() {
         when(asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS)).thenReturn(Optional.empty());
-        assertThatThrownBy(() -> detentionEngagementTeamAppealFeeDuePersonalisation.getPersonalisationForLink(asylumCase))
+
+        assertThatThrownBy(
+            () -> detentionEngagementTeamAppealFeeDuePersonalisation.getPersonalisationForLink(asylumCase))
             .isExactlyInstanceOf(IllegalStateException.class)
-            .hasMessage("internalAppealFeeIsDueLetter document not available");
+            .hasMessage("internalAppealFeeDueLetter document not available");
     }
 
     @Test
@@ -140,7 +135,7 @@ public class DetentionEngagementTeamAppealFeeDuePersonalisationTest {
         final Map<String, Object> expectedPersonalisation =
             ImmutableMap
                 .<String, Object>builder()
-                .put("subjectPrefix", detainedPrefix)
+                .put("subjectPrefix", subjectPrefix)
                 .put("appealReferenceNumber", appealReferenceNumber)
                 .put("homeOfficeReferenceNumber", homeOfficeReferenceNumber)
                 .put("appellantGivenNames", appellantGivenNames)
